add netctl support
This commit adds support for different widgets to be displayed depending
on the active netctl profile. Simply add a widgets_netctl block to your
config, and use the name of the netctl profile as the key. If there is
no key that matches your active profile, it will fall back to the normal
widgets list.